+++ CHANGELOG.md
@@ -85,6 +85,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/1360>)
 - Fix resize transform for RleMask annotation
 - (<https://github.com/openvinotoolkit/datumaro/pull/1361>)
+- Fix import YOLO variants from extractor when `urls` is not specified
+  (<https://github.com/openvinotoolkit/datumaro/pull/1362>)
 
 ## Jan. 2024 Release 1.5.2
 ### Enhancements